requirements.txt
absl-py==1.0.0
aiosignal==1.2.0
appdirs==1.4.4
astunparse==1.6.3
attrs==21.4.0
audioread==2.1.9
Automat==20.2.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
constantly==15.1.0
cycler==0.11.0
decorator==5.1.1
distlib==0.3.4
filelock==3.6.0
flatbuffers==2.0
fonttools==4.32.0
frozenlist==1.3.0
gast==0.5.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20
grpcio==1.43.0
h5py==3.6.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
incremental==21.3.0
joblib==1.1.0
jsonschema==4.4.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
librosa==0.9.1
llvmlite==0.38.0
Markdown==3.3.6
matplotlib==3.5.1
msgpack==1.0.3
neat-python==0.92
numba==0.55.1
numpy==1.21.6
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
platformdirs==2.5.2
pooch==1.6.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
pyparsing==3.0.8
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
ray==1.12.0
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.2
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
SoundFile==0.10.3.post1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.64.0
Twisted==22.2.0
typing-extensions==4.1.1
urllib3==1.26.9
virtualenv==20.14.1
Werkzeug==2.1.1
wrapt==1.14.0
zipp==3.8.0
zope.interface==5.4.0
